requirements.txt
# Core API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0  # Required for config loading
python-multipart==0.0.6

# Audio Processing
librosa==0.10.1           # Primary audio loader (Fixes Windows TorchCodec error)
soundfile==0.12.1         # Required backend for librosa
pydub==0.25.1
numpy>=1.21.0
scipy>=1.11.4

# Machine Learning & Embeddings
transformers>=4.36.2
torch>=2.1.0
torchaudio>=2.1.0         # Still required for Wav2Vec2 model definition

# Vector Database
faiss-cpu>=1.7.4

# LLM Integration
requests>=2.31.0          # CRITICAL: Added for the new LLM Client
httpx==0.26.0             # Kept for potential async compatibility
aiohttp>=3.9.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1